Project Cards v2
Now includes images.
Class names were updated to better reflect industry standards.
Created a container so space between the other components would be modified in that, instead of using the other classes.
Display was changed from "inline-block" to "flex" to make it more responsive.
CSS was updated to make the boxes more responsive in general, and to make it easier to add new boxes in the future.